+++ packages/axios-extension/src/abap/abap-service-provider.ts
@@ -8,9 +8,9 @@ import { ODataVersion } from '../base/odata-service';
 import { LayeredRepositoryService } from './lrep-service';
 import { AdtServiceName, AdtServiceConfigs, parseAtoResponse, TenantType } from './adt-catalog';
 import type { AbapServiceProviderExtension } from './abap-service-provider-extension';
-import { getTransportNumberList } from './adt-catalog/handlers/transport';
+import { getTransportRequestList } from './adt-catalog/handlers/transport';
 import { AdtCatalogService } from './adt-catalog/adt-catalog-service';
-import type { AdtCollection } from './types';
+import type { AdtCollection, TransportRequest } from './types';
 
 /**
  * Extension of the service provider for ABAP services.
@@ -187,7 +187,7 @@ export class AbapServiceProvider extends ServiceProvider implements AbapServiceP
      * @param appName Fiori project name for deployment. A new project that has not been deployed before is also allowed
      * @returns array of transports id's
      */
-    public async getTransportRequests(packageName: string, appName: string): Promise<string[]> {
+    public async getTransportRequests(packageName: string, appName: string): Promise<TransportRequest[]> {
         let serviceSchema: AdtCollection;
         try {
             serviceSchema = await this.getAdtCatalogService().getServiceDefinition(
@@ -228,6 +228,6 @@ export class AbapServiceProvider extends ServiceProvider implements AbapServiceP
             `;
 
         const response = await this.post(urlPath, data, acceptHeaders);
-        return getTransportNumberList(response.data, this.log);
+        return getTransportRequestList(response.data, this.log);
     }
 }