+++ packages/environment-check/test/cli/index.test.ts
@@ -4,11 +4,18 @@ import { cli } from '../../src/cli/index';
 import { checkEnvironment } from '../../src/checks/environment';
 import { Severity } from '../../src/types';
 import { storeResultsZip } from '../../src/output';
+import { isAppStudio } from '@sap-ux/btp-utils';
 
+jest.mock('@sap-ux/btp-utils', () => ({
+    ...(jest.requireActual('@sap-ux/btp-utils') as object),
+    isAppStudio: jest.fn()
+}));
 jest.mock('fs');
 jest.mock('../../src/checks/environment', () => ({
     checkEnvironment: jest.fn()
 }));
+
+const mockIsAppStudio = isAppStudio as jest.Mock;
 const mockCheckEnvironment = checkEnvironment as jest.Mock;
 
 jest.mock('../../src/output', () => ({
@@ -26,6 +33,7 @@ describe('Test for cli()', () => {
         // Mock setup
         console.log = jest.fn();
         process.argv = [...process.argv, '-h'];
+        mockIsAppStudio.mockReturnValue(true);
 
         // Test execution
         await cli();
@@ -58,6 +66,7 @@ describe('Test for cli()', () => {
                 ]
             })
         );
+
         console.error = jest.fn();
         console.warn = jest.fn();
         console.log = jest.fn();
@@ -80,6 +89,7 @@ describe('Test for cli()', () => {
                 messages: [{ severity: Severity.Debug, text: 'DEBUG message' }]
             })
         );
+
         console.info = jest.fn();
         process.argv = [...process.argv, '--output', 'verbose'];
 
@@ -90,7 +100,7 @@ describe('Test for cli()', () => {
         expect(console.info).toHaveBeenCalledWith('ℹ DEBUG message');
     });
 
-    test('Call cli() --destination ONE, checkEnvironment() should be called with ONE', async () => {
+    test('Call cli() --destination ONE, checkBASEnvironment() should be called with ONE', async () => {
         // Mock setup
         let checkDest;
         mockCheckEnvironment.mockImplementation((options) => {
@@ -110,7 +120,7 @@ describe('Test for cli()', () => {
         expect(checkDest).toEqual(['ONE']);
     });
 
-    test('Call cli() --destination ONE --destination TWO work/space1 work/space2, checkEnvironment() should be called with params', async () => {
+    test('Call cli() --destination ONE --destination TWO work/space1 work/space2, checkBASEnvironment() should be called with params', async () => {
         // Mock setup
         let checkDest;
         let checkWorkspace;
@@ -123,6 +133,7 @@ describe('Test for cli()', () => {
                 destinationResults: {}
             });
         });
+
         process.argv = [...process.argv, '--destination', 'ONE', '--destination', 'TWO', 'work/space1', 'work/space2'];
 
         // Test execution
@@ -143,6 +154,7 @@ describe('Test for cli()', () => {
             destinationResults: { DUMMY: { v2: 'V2DUMMY', v4: 'V4DUMMY' } }
         };
         mockCheckEnvironment.mockImplementation(() => Promise.resolve(result));
+
         jest.spyOn(mockFs, 'writeFile').mockImplementation((options, content, callback) => {
             checkWriteOtions = options;
             checkContent = content;
@@ -168,6 +180,7 @@ describe('Test for cli()', () => {
             destinationResults: { DUMMY: { v2: 'V2DUMMY', v4: 'V4DUMMY' } }
         };
         mockCheckEnvironment.mockImplementation(() => Promise.resolve(result));
+
         jest.spyOn(mockFs, 'writeFile').mockImplementation((options, content, callback) => {
             checkWriteOtions = options;
             checkContent = content;
@@ -201,6 +214,39 @@ describe('Test for cli()', () => {
         await cli();
 
         // Result check
+        expect(checkContent).toEqual(result);
+    });
+
+    test('Call cli() (VSCode)', async () => {
+        // Mock setup
+        let checkContent;
+        const result = {
+            messages: [],
+            environment: {
+                fioriGenVersion: '1',
+                cloudCli: '2',
+                appWizard: '2',
+                ui5LanguageAssistant: '2',
+                xmlToolkit: '2',
+                annotationMod: '2.2',
+                appMod: '2',
+                help: '2',
+                serviceMod: '2.4',
+                cds: '2'
+            }
+        };
+        mockIsAppStudio.mockReturnValue(false);
+        mockCheckEnvironment.mockImplementation(() => Promise.resolve(result));
+        mockStoreResultsZip.mockImplementation((content) => {
+            checkContent = content;
+        });
+        process.argv = [...process.argv, '--output', 'zip'];
+
+        // Test execution
+        await cli();
+
+        // Result check
+        expect(mockCheckEnvironment).toBeCalledTimes(1);
         expect(checkContent).toEqual(result);
     });
 });